+++ lib/nullx.dart
@@ -3,15 +3,25 @@
 /// and robust error handling, for cleaner and more resilient code
 library nullx;
 
+export 'src/also.dart';
+export 'src/bool.dart';
+export 'src/condition.dart';
+export 'src/double.dart';
 export 'src/either.dart';
 export 'src/exception.dart';
 export 'src/function.dart';
 export 'src/future.dart';
+export 'src/int.dart';
 export 'src/iterable.dart';
+export 'src/let.dart';
 export 'src/list.dart';
 export 'src/map.dart';
+export 'src/nullable.dart';
 export 'src/object.dart';
+export 'src/run.dart';
 export 'src/runes.dart';
 export 'src/stream.dart';
+export 'src/string.dart';
+export 'src/take.dart';
 export 'src/types.dart';
 export 'src/utils.dart';