+++ test/bool_test.dart
@@ -1,0 +1,28 @@
+import 'package:nullx/nullx.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('NullableBoolExtension tests', () {
+    test('orFalse returns false when boolean is null', () {
+      const bool? nullableBool = null;
+      expect(nullableBool.orFalse, isFalse);
+    });
+
+    test('orFalse returns the boolean when it is not null', () {
+      // ignore: unnecessary_nullable_for_final_variable_declarations
+      const bool? nullableBool = true;
+      expect(nullableBool.orFalse, isTrue);
+    });
+
+    test('or returns default value when boolean is null', () {
+      const bool? nullableBool = null;
+      expect(nullableBool.or(defaultValue: true), isTrue);
+    });
+
+    test('or returns the boolean when it is not null', () {
+      // ignore: unnecessary_nullable_for_final_variable_declarations
+      const bool? nullableBool = false;
+      expect(nullableBool.or(defaultValue: true), isFalse);
+    });
+  });
+}